gossip: Make boot process faster and safer
During boot up, we need to wait for gossip in multiple places:

1) Wait for gossip to settle to enable features. See commit
71bf757b2c8fb372bfd115ed00bf56069331d7cc ("gossiper: Enable features
only after gossip is settled").

2) Wait until local node knows tokens of peer nodes before announcing
join status. To make sure when the writes sent by the other nodes for
pending token range arrives, the local node knows all the tokens of the
cluster. See commit 89b769a073da1a1727b913159d85b97fe911676c.

3) Wait until peer nodes know the bootstrap tokens of local node before
streaming. To make sure all peer nodes knows the pending token range
before the streaming starts, so that all writes after streaming starts
that belong to the pending token range will be forwarded to the new node.

4) Wait until gossip is settled before start the cql server. See commit
8c909122a6a6b90ddb50ea8be250829c199616ab.

Before this patch, we sleep ring_delay which is 30 seconds by default
and wait for the size of endpoint_state_map does not change over some
time. The time out based approach does not guarantee the above
conditions are met. It can wait foo much or wait too few.

This patch introduces a new method. Instead of waiting for some
arbitrary time, we ask peer nodes to return the number of nodes it
knows, the checksum of normal tokens, the checksum of bootstrap tokens
and the checksum of gossip status (e.g., NORMAL), the checksum of nodes
leaving the cluster. If the checksums between local node and all remote
nodes are identical, it is guaranteed all nodes have the same view of
tokens and nodes status.

A new RPC verb GOSSIP_QUERY_TOKEN_STATUS is introduced for this new
method. If any of the nodes in the cluster does not support this new
verb, or a node is down and can not respond the new verb, it will
fallback to the old timeout based method. A nice thing is that we do not
need to negotiate between nodes if GOSSIP_QUERY_TOKEN_STATUS is
supported. When the node that supports GOSSIP_QUERY_TOKEN_STATUS send
this verb to node that does not support it during boot process, the
sender will get rpc::unknown_verb_error then it will fallback.

With this patch, we observed significant improvement in scylla boot time.

To bootstrap a single node with the default 30s ring_delay:
Before: 88 seconds
After: 4 seconds

To run update_cluster_layout_tests.py:TestUpdateClusterLayout.simple_add_node_1_test
Before: 141 seconds
After: 25 seconds

To run update_cluster_layout_tests.py:TestLargeScaleCluster.add_50_nodes_test
to increase a cluster from 3 nodes to 12 nodes
Before: 866 seconds
After: 111 seconds

Fixes: #4755
Refs: #4632
Tests: update_cluster_layout_tests.py dtest + manual test