storage_service: Update tokens and replace address for replace operation
If we are replacing a node with the same ip address of ourself, we
should update the tokens in token_metadata early, so that this replacing
node and peer nodes has the same hash for the normal token hash.

If we are replacing a node with the different ip address of ourself, we
should only update the new ip and tokens mapping after replace operation
is down. We can tell gossip the ip address we are replacing, so we can
skip the node being replaced in gossiper::wait_for_gossip_with_query()
which will be introduced shortly.