ldap: Allow using PasswordAuthenticator
There was a check insisting that PasswordAuthenticator be used only in
combination with CassandraRoleManager.  This was superfluous, because
CassandraRoleManager was the only manager allowed.  Now that we also
allow LDAPRoleManager, this is actually incorrect.  The check is
hereby dropped.

Fixes #1416.

Tests: unit (dev)

Signed-off-by: Dejan Mircevski <dejan@scylladb.com>
Message-Id: <20200702133420.244690-1-dejan@scylladb.com>
(cherry picked from commit f144fc9eb48852394bbebfecd54779be7da033b5)