db/view: view_updating_consumer: allow flushing the buffer mid-partition
Currently the consumer accumulates data until it has an entire
partition, then passes the resulting mutation to
`table::stream_view_replica_updates()` for further processing.  This is
problematic because if the generated mutation is too large this can
fail, which results in a deadlock, as the view update generation is
retried again, ending up generating the same overly large mutation.
To avoid this send the accumulated data for further processing based the
size of the accumulated data, not when reaching the end of the current
partition. To avoid fragmenting partitions needlessly a dual soft-hard
limit is introduced. We flush the buffer on any partition-end when above
the soft limit, or flush unconditionally when above the hard one.

(cherry picked from commit 0166f970964ee688e18a5b52f8a94a5ec8831671)