Merge "Workload prioritization: only read and create tables when appropriate" from Eliran
"
Workload prioritization used to create and use tables in the system_distributed_keyspace unconditionally.
This didn't work well in a mixed cluster because of schema issues that relate to a table that has been created in
a mixed cluster. This series introduces a WORKLOAD_PRIORITIZATION feature in order to be future proofe
and tests for the pre-existence of the WP tables to be past proof. Those together determine where it is appropriate
to create and read the WP table and are used to do so.
Another problem that this series addresses is a race condition between token initialization and reading the workload
prioritization table. This can cause and error of uninitialized token data which is not harmful (in this case) but can
alarm the user so it is better to get rid of it.

fixes #1452
fixes #1461
"

* github.com:scylladb/scylla-enterprise:
  Workload Prioritization: don't start polling the table before tokens initialization.
  workload prioritization: Only start workload prioritization when appropriate
  distributed keyspace: don't cteate workload prioritization tables if cluster doesn't support them
  features: add WORKLOAD_PRIORITIZATION feature

(cherry picked from commit 701b947692e2f34c60175f8d22038681d1fb56d7)