Scylla tables: Fix mismatched schemas with the same version
Scylla tables versions have been converted to deterministic
along with the addition of two new columns. This change was
done on master in commit f83ff8f. After merging to enterprise
this resulted in two different schemas having the same version,
the one that contains the im_memory column and the one that don't.
When causing a schema sync in a mixed cluster that contains enterprise
and non enterprise versions, this can lead to an enterprise node
trying to interpret mutations without an in_memory column using schema
that contains it resulting in an apparent corrupted mutation.
In fact the read is done in skew due to deserializing an extra value from
the middle of mutation that doesn't exist there.
This commit versions also the in_memory variants.

It is important to note that maybe a more robust solution to this
problem should be thought of since this is not the first time we
encounter problems related to different scylla_tables schemas and
it seams it is very easy to break this.

Fixes #1590

Closes #1603

(cherry picked from commit 51a3a1b70d91a049b96037b8e3c1cd07abff149f)